+++ pkg/deployment/reconcile/action.register.generated_test.go
@@ -35,12 +35,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeAddMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeAddMember.Optional())
+		})
 	})
 
 	t.Run("AppendTLSCACertificate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeAppendTLSCACertificate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeAppendTLSCACertificate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeAppendTLSCACertificate.Optional())
 		})
 	})
 
@@ -49,12 +55,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeArangoMemberUpdatePodSpec.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeArangoMemberUpdatePodSpec.Optional())
+		})
 	})
 
 	t.Run("ArangoMemberUpdatePodStatus", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeArangoMemberUpdatePodStatus)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeArangoMemberUpdatePodStatus.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeArangoMemberUpdatePodStatus.Optional())
 		})
 	})
 
@@ -63,12 +75,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeBackupRestore.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeBackupRestore.Optional())
+		})
 	})
 
 	t.Run("BackupRestoreClean", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeBackupRestoreClean)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeBackupRestoreClean.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeBackupRestoreClean.Optional())
 		})
 	})
 
@@ -77,12 +95,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeBootstrapSetPassword.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeBootstrapSetPassword.Optional())
+		})
 	})
 
 	t.Run("BootstrapUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeBootstrapUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeBootstrapUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeBootstrapUpdate.Optional())
 		})
 	})
 
@@ -91,12 +115,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeCleanOutMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeCleanOutMember.Optional())
+		})
 	})
 
 	t.Run("CleanTLSCACertificate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeCleanTLSCACertificate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeCleanTLSCACertificate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeCleanTLSCACertificate.Optional())
 		})
 	})
 
@@ -105,12 +135,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeCleanTLSKeyfileCertificate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeCleanTLSKeyfileCertificate.Optional())
+		})
 	})
 
 	t.Run("ClusterMemberCleanup", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeClusterMemberCleanup)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeClusterMemberCleanup.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeClusterMemberCleanup.Optional())
 		})
 	})
 
@@ -119,12 +155,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeDisableClusterScaling.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeDisableClusterScaling.Optional())
+		})
 	})
 
 	t.Run("DisableMaintenance", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeDisableMaintenance)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeDisableMaintenance.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeDisableMaintenance.Optional())
 		})
 	})
 
@@ -133,12 +175,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeDisableMemberMaintenance.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeDisableMemberMaintenance.Optional())
+		})
 	})
 
 	t.Run("EnableClusterScaling", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeEnableClusterScaling)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEnableClusterScaling.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEnableClusterScaling.Optional())
 		})
 	})
 
@@ -147,12 +195,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEnableMaintenance.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEnableMaintenance.Optional())
+		})
 	})
 
 	t.Run("EnableMemberMaintenance", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeEnableMemberMaintenance)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEnableMemberMaintenance.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEnableMemberMaintenance.Optional())
 		})
 	})
 
@@ -161,12 +215,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEncryptionKeyAdd.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEncryptionKeyAdd.Optional())
+		})
 	})
 
 	t.Run("EncryptionKeyPropagated", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeEncryptionKeyPropagated)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEncryptionKeyPropagated.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEncryptionKeyPropagated.Optional())
 		})
 	})
 
@@ -175,12 +235,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEncryptionKeyRefresh.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEncryptionKeyRefresh.Optional())
+		})
 	})
 
 	t.Run("EncryptionKeyRemove", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeEncryptionKeyRemove)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEncryptionKeyRemove.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEncryptionKeyRemove.Optional())
 		})
 	})
 
@@ -189,12 +255,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeEncryptionKeyStatusUpdate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeEncryptionKeyStatusUpdate.Optional())
+		})
 	})
 
 	t.Run("Idle", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeIdle)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeIdle.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeIdle.Optional())
 		})
 	})
 
@@ -203,12 +275,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTAdd.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTAdd.Optional())
+		})
 	})
 
 	t.Run("JWTClean", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeJWTClean)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTClean.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTClean.Optional())
 		})
 	})
 
@@ -217,12 +295,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTPropagated.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTPropagated.Optional())
+		})
 	})
 
 	t.Run("JWTRefresh", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeJWTRefresh)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTRefresh.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTRefresh.Optional())
 		})
 	})
 
@@ -231,12 +315,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTSetActive.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTSetActive.Optional())
+		})
 	})
 
 	t.Run("JWTStatusUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeJWTStatusUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeJWTStatusUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeJWTStatusUpdate.Optional())
 		})
 	})
 
@@ -245,12 +335,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeKillMemberPod.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeKillMemberPod.Optional())
+		})
 	})
 
 	t.Run("LicenseSet", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeLicenseSet)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeLicenseSet.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeLicenseSet.Optional())
 		})
 	})
 
@@ -259,12 +355,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeMarkToRemoveMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeMarkToRemoveMember.Optional())
+		})
 	})
 
 	t.Run("MemberPhaseUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeMemberPhaseUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeMemberPhaseUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeMemberPhaseUpdate.Optional())
 		})
 	})
 
@@ -273,12 +375,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeMemberRIDUpdate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeMemberRIDUpdate.Optional())
+		})
 	})
 
 	t.Run("PVCResize", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypePVCResize)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypePVCResize.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypePVCResize.Optional())
 		})
 	})
 
@@ -287,12 +395,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypePVCResized.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypePVCResized.Optional())
+		})
 	})
 
 	t.Run("PlaceHolder", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypePlaceHolder)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypePlaceHolder.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypePlaceHolder.Optional())
 		})
 	})
 
@@ -301,12 +415,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRebalancerCheck.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRebalancerCheck.Optional())
+		})
 	})
 
 	t.Run("RebalancerClean", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRebalancerClean)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRebalancerClean.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRebalancerClean.Optional())
 		})
 	})
 
@@ -315,12 +435,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.True(t, api.ActionTypeRebalancerGenerate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRebalancerGenerate.Optional())
+		})
 	})
 
 	t.Run("RecreateMember", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRecreateMember)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRecreateMember.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRecreateMember.Optional())
 		})
 	})
 
@@ -329,12 +455,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRefreshTLSKeyfileCertificate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRefreshTLSKeyfileCertificate.Optional())
+		})
 	})
 
 	t.Run("RemoveMember", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRemoveMember)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRemoveMember.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRemoveMember.Optional())
 		})
 	})
 
@@ -343,12 +475,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRenewTLSCACertificate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRenewTLSCACertificate.Optional())
+		})
 	})
 
 	t.Run("RenewTLSCertificate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRenewTLSCertificate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRenewTLSCertificate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRenewTLSCertificate.Optional())
 		})
 	})
 
@@ -357,12 +495,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeResignLeadership.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeResignLeadership.Optional())
+		})
 	})
 
 	t.Run("ResourceSync", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeResourceSync)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeResourceSync.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeResourceSync.Optional())
 		})
 	})
 
@@ -372,6 +516,9 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRotateMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRotateMember.Optional())
+		})
 	})
 
 	t.Run("RotateStartMember", func(t *testing.T) {
@@ -380,12 +527,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRotateStartMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRotateStartMember.Optional())
+		})
 	})
 
 	t.Run("RotateStopMember", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRotateStopMember)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRotateStopMember.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRotateStopMember.Optional())
 		})
 	})
 
@@ -394,12 +547,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRuntimeContainerArgsLogLevelUpdate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRuntimeContainerArgsLogLevelUpdate.Optional())
+		})
 	})
 
 	t.Run("RuntimeContainerImageUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeRuntimeContainerImageUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRuntimeContainerImageUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRuntimeContainerImageUpdate.Optional())
 		})
 	})
 
@@ -408,12 +567,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeRuntimeContainerSyncTolerations.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRuntimeContainerSyncTolerations.Optional())
+		})
 	})
 
 	t.Run("SetCondition", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeSetCondition)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetCondition.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetCondition.Optional())
 		})
 	})
 
@@ -422,12 +587,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetConditionV2.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetConditionV2.Optional())
+		})
 	})
 
 	t.Run("SetCurrentImage", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeSetCurrentImage)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetCurrentImage.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetCurrentImage.Optional())
 		})
 	})
 
@@ -436,12 +607,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetCurrentMemberArch.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetCurrentMemberArch.Optional())
+		})
 	})
 
 	t.Run("SetMaintenanceCondition", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeSetMaintenanceCondition)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetMaintenanceCondition.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetMaintenanceCondition.Optional())
 		})
 	})
 
@@ -450,6 +627,9 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetMemberCondition.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetMemberCondition.Optional())
+		})
 	})
 
 	t.Run("SetMemberConditionV2", func(t *testing.T) {
@@ -457,12 +637,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetMemberConditionV2.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetMemberConditionV2.Optional())
+		})
 	})
 
 	t.Run("SetMemberCurrentImage", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeSetMemberCurrentImage)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeSetMemberCurrentImage.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeSetMemberCurrentImage.Optional())
 		})
 	})
 
@@ -472,12 +658,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeShutdownMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeShutdownMember.Optional())
+		})
 	})
 
 	t.Run("TLSKeyStatusUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeTLSKeyStatusUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTLSKeyStatusUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTLSKeyStatusUpdate.Optional())
 		})
 	})
 
@@ -486,12 +678,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTLSPropagated.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTLSPropagated.Optional())
+		})
 	})
 
 	t.Run("TimezoneSecretSet", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeTimezoneSecretSet)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTimezoneSecretSet.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTimezoneSecretSet.Optional())
 		})
 	})
 
@@ -500,12 +698,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTopologyDisable.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTopologyDisable.Optional())
+		})
 	})
 
 	t.Run("TopologyEnable", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeTopologyEnable)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTopologyEnable.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTopologyEnable.Optional())
 		})
 	})
 
@@ -514,12 +718,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTopologyMemberAssignment.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTopologyMemberAssignment.Optional())
+		})
 	})
 
 	t.Run("TopologyZonesUpdate", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeTopologyZonesUpdate)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeTopologyZonesUpdate.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeTopologyZonesUpdate.Optional())
 		})
 	})
 
@@ -528,12 +738,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeUpToDateUpdate.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeUpToDateUpdate.Optional())
+		})
 	})
 
 	t.Run("UpdateTLSSNI", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeUpdateTLSSNI)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeUpdateTLSSNI.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeUpdateTLSSNI.Optional())
 		})
 	})
 
@@ -542,12 +758,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeUpgradeMember.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeUpgradeMember.Optional())
+		})
 	})
 
 	t.Run("WaitForMemberInSync", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeWaitForMemberInSync)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeWaitForMemberInSync.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeWaitForMemberInSync.Optional())
 		})
 	})
 
@@ -556,12 +778,18 @@ func Test_Actions(t *testing.T) {
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeWaitForMemberReady.Internal())
 		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeWaitForMemberReady.Optional())
+		})
 	})
 
 	t.Run("WaitForMemberUp", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeWaitForMemberUp)
 		t.Run("Internal", func(t *testing.T) {
 			require.False(t, api.ActionTypeWaitForMemberUp.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeWaitForMemberUp.Optional())
 		})
 	})
 }