+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Maintenance) Fix generated license dates
 - (Improvement) Reduce CI on Commit Travis runs
 - (Maintenance) Add license range rewrite command
+- (Feature) Optional Action
 
 ## [1.2.24](https://github.com/arangodb/kube-arangodb/tree/1.2.24) (2023-01-25)
 - (Bugfix) Fix deployment creation on ARM64